requirements.txt
gradio
transformers
torch
accelerate
pdfplumber
pytesseract
Pillow
